requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
mysqlclient
alembic
passlib[bcrypt]
pyjwt
cachetools
python-jose
